Fixed the popup trigger Button
The button kept moving around after you clicked it. Now it is centered at the bottom of the page. IT still moves an mm up after being clicked but then stays. Also fixed the width of the pop up window. Need to fix the width of the input box for the description.